+++ proton-mail-uniffi/src/mail/datatypes/available_action.rs
@@ -1,11 +1,15 @@
 mod conversation;
 mod general_actions;
+mod is_selected;
+mod label_as_action;
 mod message;
 mod move_action;
 mod reply_action;
 
 pub use self::conversation::*;
 pub use self::general_actions::*;
+pub use self::is_selected::*;
+pub use self::label_as_action::*;
 pub use self::message::*;
 pub use self::move_action::*;
 pub use self::reply_action::*;